pyproject.toml
[tool.poetry]
name = "clone-client"
version = "0.0.42"
description = ""
authors = ["Clone Inc."]
readme = "README.md"
packages = [{include = "clone_client"}]
include = [
    { path = "clone_client/proto/*", format = ["sdist","wheel"] }
]

[tool.poetry.dependencies]
pydantic = {version = "2.8.2", extras = ["dotenv"]}
python = "^3.10"
zeroconf = "^0.132.2"
grpcio = "^1.65.5"
grpcio-tools = "^1.65.5"
pydantic-settings = "^2.4.0"
protobuf = "^5.29.0"
numpy = "2.2"
scipy = "1.15.3"

[tool.poetry.group.dev.dependencies]
bandit = "^1.7.9"
black = "^24.8.0"
flake8 = "^7.1.1"
flake8-comprehensions = "^3.15.0"
flake8-isort = "^6.1.1"
mypy = "1.11.1"
pylint = "^3.2.6"
taskipy = "^1.13.0"
mypy-protobuf = "^3.6.0"
types-protobuf = "^5.29.0"
pylint-pydantic = "^0.3.2"
pylint-protobuf = "^0.22.0"
pytest = "^8.3.2"
types-psutil = "^6.0.0.20240621"
debugpy = "^1.8.11"
dearpygui = "^2.1.0"

[[tool.poetry.source]]
name = "clone-pypi"
url = "https://forgejo.cloneai.tech/api/packages/clone/pypi/simple"
priority = "supplemental"

[tool.taskipy.tasks]
fmt = "isort ./clone_client && black ./clone_client"
lint = """
    black --check ./clone_client && \
    pylint ./clone_client --prefer-stubs y && \
    flake8 ./ && \
    bandit -r ./ -x ./.venv -ll
"""
check-types = "mypy ./"
run-protoc = "python ./run_protoc.py"

[tool.black]
line-length = 110
target-version = ['py311']
exclude = 'pb2.pyi?$|pb2_grpc.py$'

[tool.isort]
profile = "black"
multi_line_output = 3
force_alphabetical_sort_within_sections = true
force_sort_within_sections = true
skip_glob = ["**/*_pb2*[@(py\\0|pyi]"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
